fix: Update instruction evaluators to correct type conversions for D2L and L2F

+++ core/src/main/java/tokyo/peya/langjal/compiler/instructions/cast/InstructionEvaluatorD2L.java
@@ -20,7 +20,7 @@ public class InstructionEvaluatorD2L extends AbstractSingleInstructionEvaluator<
     {
         return FrameDifferenceInfo.builder(instruction)
                                   .popPrimitive(StackElementType.DOUBLE)
-                                  .pushPrimitive(StackElementType.INTEGER)
+                                  .pushPrimitive(StackElementType.LONG)
                                   .build();
     }
