+++ core/src/main/java/tokyo/peya/langjal/compiler/instructions/cast/InstructionEvaluatorL2F.java
@@ -20,7 +20,7 @@ public class InstructionEvaluatorL2F extends AbstractSingleInstructionEvaluator<
     {
         return FrameDifferenceInfo.builder(instruction)
                                   .popPrimitive(StackElementType.LONG)
-                                  .pushPrimitive(StackElementType.DOUBLE)
+                                  .pushPrimitive(StackElementType.FLOAT)
                                   .build();
     }
